_gsdocs-proposals/requirements.txt
fastapi==0.115.9
uvicorn==0.29.0
python-multipart==0.0.9

torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

transformers==4.41.0
diffusers==0.27.2
accelerate==0.28.0
safetensors==0.4.3

pillow==10.3.0
numpy==1.26.4
scipy==1.12.0
opencv-python-headless==4.10.0.84

realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8

python-dotenv==1.0.1
huggingface_hub==0.33.4
tokenizers==0.19.1